draw custom mini-map

+++ ray_casting/ray_casting.c
@@ -6,7 +6,7 @@
 /*   By: del-yaag <del-yaag@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/13 14:30:05 by del-yaag          #+#    #+#             */
-/*   Updated: 2023/08/16 23:46:36 by del-yaag         ###   ########.fr       */
+/*   Updated: 2023/08/18 20:43:51 by del-yaag         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -73,7 +73,7 @@ double	ver_ray_casting(t_cub *data)
 	vertical_ray_equation(data);
 	while (data->cast.vx > 0 && data->cast.vy > 0
 		&& data->cast.vy < data->cordt.height
-		&& (size_t)(data->cast.vx / UNIT) < \
+		&& (data->cast.vx / UNIT) < \
 		ft_strlen(data->draw.line[(int)(data->cast.vy / UNIT)]))
 	{
 		if (data->draw.line[(int)(data->cast.vy / UNIT)] \
@@ -99,7 +99,7 @@ double	hor_ray_casting(t_cub *data)
 	horizontal_ray_equation(data);
 	while (data->cast.hx >= 0 && data->cast.hy >= 0
 		&& data->cast.hy < data->cordt.height
-		&& (size_t)(data->cast.hx / UNIT) < \
+		&& (data->cast.hx / UNIT) < \
 		ft_strlen(data->draw.line[(int)(data->cast.hy / UNIT)]))
 	{
 		if (data->draw.line[(int)(data->cast.hy / UNIT)] \
